+++ CHANGES.md
@@ -16,6 +16,7 @@
 * Exposed the `force_retain_solver` parameter in `cysolve_ivp`.
   * If set to True (the default) then the solver will be retained at the end of integration, even if it is not necessarily needed. This can avoid issues where the solver is accessed by the user after it has been released. 
   * If set to False (and capture_extra = False or capture_dense = False) then the solver's memory will be released at the end of integration. This can reduce memory overhead particularly if your problem has a large number of dependent y variables.
+* Changed how the RK "K" matrix is stored in memory so it is better optimized to cache lines.
 
 ##### Benchmarks
 * Added in a repeater inside `cysolve_ivp` tester so we can more accurately benchmark the integrators performance without also tracking the python wrappers overhead.